requirements.txt
streamlit
torch
torchvision
pandas
numpy
scikit-learn
matplotlib
shap
Pillow
pydeck
plotly
geopy